LPS-196088 Inline UpgradeStepRegistratorTracker into UpgradeExecutor
