+++ modules/apps/portal/portal-upgrade-impl/src/main/java/com/liferay/portal/upgrade/internal/executor/UpgradeExecutor.java
@@ -5,35 +5,58 @@
 
 package com.liferay.portal.upgrade.internal.executor;
 
+import com.liferay.osgi.util.ServiceTrackerFactory;
+import com.liferay.petra.lang.SafeCloseable;
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.db.index.IndexUpdaterUtil;
 import com.liferay.portal.kernel.cache.CacheRegistryUtil;
+import com.liferay.portal.kernel.configuration.Configuration;
+import com.liferay.portal.kernel.configuration.ConfigurationFactoryUtil;
+import com.liferay.portal.kernel.dao.jdbc.DataAccess;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Release;
 import com.liferay.portal.kernel.model.ReleaseConstants;
 import com.liferay.portal.kernel.module.util.BundleUtil;
 import com.liferay.portal.kernel.service.ReleaseLocalService;
+import com.liferay.portal.kernel.upgrade.UpgradeException;
 import com.liferay.portal.kernel.upgrade.UpgradeStep;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.version.Version;
+import com.liferay.portal.tools.DBUpgrader;
+import com.liferay.portal.upgrade.PortalUpgradeProcess;
 import com.liferay.portal.upgrade.internal.graph.ReleaseGraphManager;
 import com.liferay.portal.upgrade.internal.registry.UpgradeInfo;
-import com.liferay.portal.upgrade.internal.registry.UpgradeStepRegistratorTracker;
+import com.liferay.portal.upgrade.internal.registry.UpgradeStepRegistry;
 import com.liferay.portal.upgrade.internal.release.ReleasePublisher;
 import com.liferay.portal.upgrade.log.UpgradeLogContext;
+import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
 
+import java.sql.Connection;
+import java.sql.SQLException;
+
+import java.util.ArrayList;
 import java.util.Dictionary;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Properties;
 
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
+import org.osgi.framework.FrameworkUtil;
+import org.osgi.framework.ServiceReference;
+import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
+import org.osgi.util.tracker.ServiceTracker;
+import org.osgi.util.tracker.ServiceTrackerCustomizer;
 
 /**
  * @author Preston Crary
@@ -47,8 +70,7 @@ public class UpgradeExecutor {
 
 		String schemaVersionString = "0.0.0";
 
-		Release release = _upgradeStepRegistratorTracker.fetchUpgradedRelease(
-			bundle.getSymbolicName());
+		Release release = _fetchUpgradedRelease(bundle.getSymbolicName());
 
 		if ((release != null) &&
 			Validator.isNotNull(release.getSchemaVersion())) {
@@ -137,15 +159,24 @@ public class UpgradeExecutor {
 
 	@Activate
 	protected void activate(BundleContext bundleContext) {
-		_upgradeStepRegistratorTracker = new UpgradeStepRegistratorTracker(
-			bundleContext, _releaseLocalService, this);
+		_bundleContext = bundleContext;
 
-		_upgradeStepRegistratorTracker.open();
+		try (Connection connection = DataAccess.getConnection()) {
+			_portalUpgraded = PortalUpgradeProcess.isInLatestSchemaVersion(
+				connection);
+		}
+		catch (SQLException sqlException) {
+			throw new RuntimeException(sqlException);
+		}
+
+		_serviceTracker = ServiceTrackerFactory.open(
+			bundleContext, UpgradeStepRegistrator.class,
+			new UpgradeStepRegistratorServiceTrackerCustomizer());
 	}
 
 	@Deactivate
 	protected void deactivate() {
-		_upgradeStepRegistratorTracker.close();
+		_serviceTracker.close();
 	}
 
 	private void _executeUpgradeInfos(
@@ -204,6 +235,35 @@ public class UpgradeExecutor {
 		CacheRegistryUtil.clear();
 	}
 
+	private Release _fetchUpgradedRelease(String bundleSymbolicName) {
+		Release release = _releaseLocalService.fetchRelease(bundleSymbolicName);
+
+		if (release == null) {
+			List<UpgradeStep> releaseUpgradeSteps = _releaseUpgradeStepsMap.get(
+				bundleSymbolicName);
+
+			if (releaseUpgradeSteps != null) {
+				for (UpgradeStep releaseUpgradeStep : releaseUpgradeSteps) {
+					try {
+						UpgradeLogContext.setContext(bundleSymbolicName);
+
+						releaseUpgradeStep.upgrade();
+					}
+					catch (UpgradeException upgradeException) {
+						_log.error(upgradeException);
+					}
+					finally {
+						UpgradeLogContext.clearContext();
+					}
+				}
+
+				release = _releaseLocalService.fetchRelease(bundleSymbolicName);
+			}
+		}
+
+		return release;
+	}
+
 	private boolean _isInitialRelease(List<UpgradeInfo> upgradeInfos) {
 		UpgradeInfo upgradeInfo = upgradeInfos.get(0);
 
@@ -245,12 +305,135 @@ public class UpgradeExecutor {
 	private static final Log _log = LogFactoryUtil.getLog(
 		UpgradeExecutor.class);
 
+	private BundleContext _bundleContext;
+	private boolean _portalUpgraded;
+
 	@Reference
 	private ReleaseLocalService _releaseLocalService;
 
 	@Reference
 	private ReleasePublisher _releasePublisher;
 
-	private UpgradeStepRegistratorTracker _upgradeStepRegistratorTracker;
+	private final Map<String, List<UpgradeStep>> _releaseUpgradeStepsMap =
+		new HashMap<>();
+	private ServiceTracker<UpgradeStepRegistrator, SafeCloseable>
+		_serviceTracker;
+
+	private class UpgradeStepRegistratorServiceTrackerCustomizer
+		implements ServiceTrackerCustomizer
+			<UpgradeStepRegistrator, SafeCloseable> {
+
+		@Override
+		public SafeCloseable addingService(
+			ServiceReference<UpgradeStepRegistrator> serviceReference) {
+
+			UpgradeStepRegistrator upgradeStepRegistrator =
+				_bundleContext.getService(serviceReference);
+
+			if (upgradeStepRegistrator == null) {
+				return null;
+			}
+
+			Class<? extends UpgradeStepRegistrator> clazz =
+				upgradeStepRegistrator.getClass();
+
+			Bundle bundle = FrameworkUtil.getBundle(clazz);
+
+			String bundleSymbolicName = bundle.getSymbolicName();
+
+			int buildNumber = 0;
+
+			ClassLoader classLoader = clazz.getClassLoader();
+
+			if (classLoader.getResource("service.properties") != null) {
+				Configuration configuration =
+					ConfigurationFactoryUtil.getConfiguration(
+						classLoader, "service");
+
+				Properties properties = configuration.getProperties();
+
+				buildNumber = GetterUtil.getInteger(
+					properties.getProperty("build.number"));
+			}
+
+			UpgradeStepRegistry upgradeStepRegistry = new UpgradeStepRegistry(
+				buildNumber);
+
+			upgradeStepRegistrator.register(upgradeStepRegistry);
+
+			List<UpgradeStep> releaseUpgradeSteps =
+				upgradeStepRegistry.getReleaseCreationUpgradeSteps();
+
+			if (!releaseUpgradeSteps.isEmpty()) {
+				_releaseUpgradeStepsMap.put(
+					bundleSymbolicName, releaseUpgradeSteps);
+			}
+
+			List<UpgradeInfo> upgradeInfos =
+				upgradeStepRegistry.getUpgradeInfos(_portalUpgraded);
+
+			if (DBUpgrader.isUpgradeDatabaseAutoRunEnabled() ||
+				(_releaseLocalService.fetchRelease(bundleSymbolicName) ==
+					null)) {
+
+				try {
+					execute(bundle, upgradeInfos);
+				}
+				catch (Throwable throwable) {
+					_log.error(
+						"Failed upgrade process for module ".concat(
+							bundleSymbolicName),
+						throwable);
+				}
+			}
+
+			List<ServiceRegistration<UpgradeStep>> serviceRegistrations =
+				new ArrayList<>(upgradeInfos.size());
+
+			for (UpgradeInfo upgradeInfo : upgradeInfos) {
+				ServiceRegistration<UpgradeStep> serviceRegistration =
+					_bundleContext.registerService(
+						UpgradeStep.class, upgradeInfo.getUpgradeStep(),
+						HashMapDictionaryBuilder.<String, Object>put(
+							"build.number", upgradeInfo.getBuildNumber()
+						).put(
+							"upgrade.bundle.symbolic.name", bundleSymbolicName
+						).put(
+							"upgrade.from.schema.version",
+							upgradeInfo.getFromSchemaVersionString()
+						).put(
+							"upgrade.to.schema.version",
+							upgradeInfo.getToSchemaVersionString()
+						).build());
+
+				serviceRegistrations.add(serviceRegistration);
+			}
+
+			return () -> {
+				for (ServiceRegistration<UpgradeStep> serviceRegistration :
+						serviceRegistrations) {
+
+					serviceRegistration.unregister();
+				}
+
+				_releaseUpgradeStepsMap.remove(bundleSymbolicName);
+			};
+		}
+
+		@Override
+		public void modifiedService(
+			ServiceReference<UpgradeStepRegistrator> serviceReference,
+			SafeCloseable safeCloseable) {
+		}
+
+		@Override
+		public void removedService(
+			ServiceReference<UpgradeStepRegistrator> serviceReference,
+			SafeCloseable safeCloseable) {
+
+			safeCloseable.close();
+		}
+
+	}
 
 }
