+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/check/JavaVerifyUpgradeConnectionCheck.java
@@ -25,7 +25,7 @@ public class JavaVerifyUpgradeConnectionCheck extends BaseFileCheck {
 			fileName.endsWith("DBUpgrader.java") ||
 			fileName.endsWith("UpgradeRecorder.java") ||
 			fileName.endsWith("Test.java") ||
-			fileName.endsWith("UpgradeStepRegistratorTracker.java") ||
+			fileName.endsWith("UpgradeExecutor.java") ||
 			fileName.endsWith("UpgradeTableListener.java") ||
 			content.contains("Callable<Void>")) {
